+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Control over client visibility of entities.
+
 ## [0.20.0] - 2024-01-13
 
 ### Fixed