+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2024-01-22
+
 ### Added
 
 - Control over client visibility of entities.
@@ -339,7 +341,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release after separation from [Project Harmonia](https://github.com/projectharmonia/project_harmonia).
 
-[unreleased]: https://github.com/projectharmonia/bevy_replicon/compare/v0.20.0...HEAD
+[unreleased]: https://github.com/projectharmonia/bevy_replicon/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/projectharmonia/bevy_replicon/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/projectharmonia/bevy_replicon/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/projectharmonia/bevy_replicon/compare/v0.18.2...v0.19.0
 [0.18.2]: https://github.com/projectharmonia/bevy_replicon/compare/v0.18.1...v0.18.2